A bit more renaming/refactoring
